requirements.txt
asgiref==3.8.1
attrs==25.3.0
django==4.2.21
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
pip==25.1.1
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.25.1
setuptools==80.9.0
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
